Rename engine to loop#collect
