+++ lib/snowman-io/launcher.rb
@@ -8,12 +8,12 @@ module SnowmanIO
 
     def start
       @web_server_supervisor = WebServer.supervise_as(:web_server, API, @options.slice(:port, :host, :verbose))
-      @engine = Engine.new_link
-      @engine.async.tick
+      @collect = Loop::Collect.new_link
+      @collect.async.tick
     end
 
     def stop
-      @engine.terminate
+      @collect.terminate
       @web_server_supervisor.terminate # TODO: shutdown blocking?
     end
   end
